rewrite assertion response, change get_json_function
